Add support to store also warmup times
With this commit we prepare the metrics store to store also warmup
samples. As Rally is nearing a public release it is easier to add
this now instead of adding changes later.

Note that we still only store normal samples. Support for also
storing warmup samples will come in #80.

Relates #80