Record race meta-data and allow to list races
With this commit we record every race in the metrics store. This
allows us also to list them afterwards. This is a preparation
step for the tournament mode which will allow users to compare the
results of two races.

We also show all output of the list command in tabular form.

Relates #57