Switch to process per client instead of thread per client
With this commit we change the client model of one process per
client instead of one thread per client. We also allow to run
queries by more than one client.

Clients communicate internally via an actor system so we are
already preparing Rally for truly distributed benchmarks which
will be implemented in #71.

Closes #58
Closes #64
Closes #108