Make PartitionBulkIndexParamSource more flexible
With this commit, PartitionBulkIndexParamSource takes two
additional parameters `batch-size` and `action-and-meta-data`.

`batch-size` allows to read a multiple of `bulk-size` lines
from the source file. This is intended to reduce the overhead of
reading for small bulk sizes. By default, `batch-size` is set
equal to `bulk-size`.

`action-and-meta-data` allows to specify how to treat the action and
metadata line for bulk indexing. There are three possible values:

* `generate` (the default), means that the action and metadata line
  will be generated by Rally.
* `none` means that no action and metadata line should be generated.
  Rally will then use the provided index and type to tell
  Elasticsearch for which index and type documents should be created.
* `sourcefile` means that Rally will assume that the source file
  contains an explicit action and metadata line for each document and
  will send it as is to Elasticsearch. This allows users to specify
  e.g. arbitrary document ids but also to delete or update documents,
  not just index them.

Closes #111