Make Rally work in degraded mode when offline
With this commit we consistently switch to a queue-based implementation
of Rally internal's actor system when the (default) TCP-based
implementation cannot be initialized (this happens when the machine is
offline).

Rally will continue to work as long it is not necessary to distribute
it across multiple machines (which requires a TCP-based implementation).
This will most likely be a problem in corporate environments and in
cases where Rally has to work distributed.

Closes #261