Make 'race' self-contained
With this commit, we establish 'race' as a self-contained entity. As
this touches several areas in the metric store implementation, we
use this opportunity to implement a couple of breaking changes at once.
Namely:

* The least supported metrics store version is now Elasticsearch 5.0
* Race and metrics are now stored in individual indices (rally-metrics-*
and rally-races-* respectively).
* New indices are created once per month instead of once per year.

Furthermore, the `list facts` subcommand is deprecated and will be
removed soon as Rally gathers all relevant information now itself and
stores it in the race.

We also implement a file-based version of the race store, so users can
now use the tournament feature without configuring a dedicated
Elasticsearch metrics store. This also simplifies sharing of race
results, as users just need to share the race results file.

Closes #138
Closes #220
Closes #258
Closes #279
Relates #282