Increase launcher's process wait timeout
With this commit we increase the internal launcher's process wait
timeout from 60 to 90. The reason is that we have seen failing builds
due to stalls at Elasticsearch startup that are coming from pretouching
the Java heap (because of `-XX:+AlwaysPreTouch`). An alternative would
be to drop caches with

```
echo 3 > /proc/sys/vm/drop_caches
sync
```

but this is not feasible on our build servers (see also
https://access.redhat.com/solutions/2685771).