Reset relative time per task
With this commit we reset the relative time in all metrics store
instances (driver: request metrics, mechanic: system metrics). The way
it is implemented there may be a small time lag (in the range of
single-digit milliseconds) between request metrics and system metrics.
We may be able to reduce this time lag by measuring round-trip times
between actors but this would complicate the implementation considerably
and might not be effective anyway.

Closes #278