Autodetect and use Java 9
With this commit we auto-detect (and prefer) a Java 9 GA release in the
initial configuration routine. We also change the name of the config
property `java8.home` to the more generic name `java.home`.

Java 9 has added a module system and Gradle requires some special
permissions. We add the necessary permissions when we use a JDK > 8.

Testing has also shown that ES 1.x fails on Java 9 due to problems with
bootclasspath. As ES 1.x went EOL several months ago, we remove it from
our integration test matrix.

Closes #265