Only map numeric HTTP status codes
We have previously checked for the pseudo status code "N/A" when
handling transport errors but the Python Elasticsearch clients also
returns other strings (like "TIMEOUT"). This is problematic as Rally
requires it to be an int when writing it to the metrics store.

With this commit we check whether the status code is a proper int and
only then set it.

Closes #337