Bulk index all data by default
With this commit we also query the parameter source when determining the
default number of iterations. Previously, when the user did not specify
any time-period nor any number of iterations we always defaulted to zero
warmup iterations and one measurement iteration. This lead to surprising
behavior for bulk-indexing when the user forgot to add a warmup time
period because we only issued one bulk request.

Closes #377
Relates #436