Bump pytest* versions and beautify make test output
Take advantage of the colored and verbose output options of newer
versions of pytest to make the output of `make test`/`make
coverage`/`make benchmark` easier to follow.

Also bump pytest to the latest version to support progress
percentage[1]. While at it, also bump pytest-benchmark to the latest
version.

[1]
https://docs.pytest.org/en/latest/changelog.html#pytest-3-3-0-2017-11-23

Relates  #437