Require java-10 to build Elasticsearch from source
Similar to the work done for #387, following the merge of the
Elasticsearch PR[1], Rally needs to require java-10 to build
Elasticsearch from source.

Bump the configuration version to 14 and adjust migration of existing
configs to support building with java-10. Also add needed test cases
and update docs.

Note that building from source with java-10 requires, at minimum,
gradle 4.5[2]. This helps us remove a number of additional gradle
options that were needed for java-9. So for now, building will fail
with gradle <4.5, which will won't be a worry at all once we start
using `./gradlew`; this is tracked in issue#412.

Closes #450
Relates #451 

[1] https://github.com/elastic/elasticsearch/pull/29174
[2] https://github.com/gradle/gradle/pull/3892