Improve simulation of bulk-indexing conflicts
* Improve simulation of bulk-indexing conflicts

With this commit we introduce a new property `conflict-probability` for
the bulk-indexing parameter source. Previously we had a hard-codded
probability of 25% but now the user can control it.

We also add a new property `on-conflict` which allows users
to define whether the action-and-metadata line should use "index" or
"update" on simulated id conflicts (the default is "index").

Closes #422