Create unique file name in race store
With this commit we use the user-tag `name` as suffix for the race store
file name in case the user has provided an effective start date. If that
is not present, we use the internal race id as a fallback.

Previously we have used the combination of track, challenge and car as a
suffix but as these can be parameterized (car-params, track-params),
this combination is not necessarily unique anymore.