Don't log all files in the data directory
With this commit we do not log the files in the data directory anymore.
This causes problems due to producing too many log entries (one log
entry per line) which overwhelms the actor system (every log message is
forwarded to a central logging process). Besides, we do not need this
information and it was only included for historical purposes. Third,
this mechanism was broken as well in the distributed case (because
logging was partially broken before #503) and thus it did not cause
problems so far.

Relates #503