Resolve artefact name based on download URL
With this commit we resolve the artefact name based on the download URL
instead of creating a static name ourselves. This avoids situations
where there are two variations of a distribution (default and OSS) with
differently named artefacts which Rally previously resolved to the same
file name and mistakenly picked whichever has been downloaded first.

Closes #525
Relates #526