Defer startup of mechanic actor
With this commit we start the mechanic actor as late as possible in
order to avoid race conditions if errors occur early during
initialization. If the actor is created while an error is being thrown
it can happen that it misses the message to terminate and thus the
process will linger.

Closes #561
Relates #562