Always use remote branch for updates
With this commit we always attempt to checkout the remote branch first.
In a scenario where a user has already cloned the repository but uses
the `--offline` flag the correct branches might not be proper local
tracking branches but are already available on the target machine as
remote branches.

Closes #573
Relates #575