Temporarily disable Python 3.4 in build
With this commit we temporarily remove Python 3.4 testing from our
builds. Due an OS upgrade in the CI matrix, Python 3.4 is not easily
available anymore so we will build only with 3.5, 3.6 and 3.7 for the
time being. We will decide at a later point whether we will drop testing
on 3.4 completely or we use a different solution for testing with Python
3.4

Relates #641