Allow passing any request parameter to the create index API
https://github.com/elastic/rally-tracks/pull/64 requires passing a
parameter in the create-index runner, however, the underlying ES Python
client method[1] doesn't support it. Additionally our docs[2] hint
that this is possible, but it isn't.

This isn't caught by unit tests as the ES client is mocked and an
integration test is needed (TODO in a separate PR).

Directly call es.transport.perform_request to create an index, allowing
passing of any arbitrary parameter (such as `include_type_name`) to
the create index Elasticsearch API.

[1] https://elasticsearch-py.readthedocs.io/en/master/api.html?highlight=indices%20create#elasticsearch.client.IndicesClient.create
[2] https://esrally.readthedocs.io/en/stable/track.html#create-index

Relates #642