Honor JAVA_HOME in provisioner
We currently rely on `java` being available on the PATH when
provisioning an Elasticsearch node but instead we should be consistent
and always use the specified runtime JDK when provisoning a node. With
this commit we determine the correct runtime JDK and set it as
`JAVA_HOME` when provisioning a node.

Relates #673