Remove node_count_per_host
With this commit we remove the provisioner variable
`node_count_per_host` which has been used to define the setting
`node.max_local_storage_nodes` in Elasticsearch. As the only value that
we ever allow (`1`) is identical to Elasticsearch's default, we remove
support for this variable.

Relates elastic/rally-teams#27
Relates #691