Always exit on OOME
With this commit we add the JVM flag `ExitOnOutOfMemoryError`
unconditionally when Elasticsearch is configured by Rally. Previously we
had to check whether the JVM in use supports this flag because we
supported Elasticsearch 1.x which can be run with Java 1.7. As the JVM flag
has only been introduced with Java 8, we had a check in place. Now that
we have dropped support for Elasticsearch 1.x (in #716), we can safely
assume that the JVM supports this flag and unconditionally set it.

Relates #715
Relates #723