Log git output
With this commit we send git output to logs (except when cloning a repo
initially as it might require the user to enter credentials). This
avoids that git output interferes with Rally's. This is needed e.g. when
Rally's `download` subcommand is invoked where we usually expect only a
JSON fragment as output.

Closes #745
Relates #747