Check that origin points to main repo for release
* Check that origin points to main repo for release

With this commit we add a new pre-release check that ensures that the
git remote named `origin` points to Rally's main repository instead of a
fork. This is necessary for releases because later during the release
process, changes will be pushed automatically to the main repository
(such as the release tags).

Relates #762
Closes #717