Run a task completely even without time-periods
There are multiple conditions how a task can be completed. Either the user
explicitly specifies a maximum number of iterations or a time-period or a task
is implicitly completed when the parameter source is exhausted (e.g. when
reading parameters from a file). In the latter case it is not necessary to
provide any indication when a task is complete because the parameter source can
determine it. However, so far this was required.

With this commit we introduce a change in the parameter source API so that it is
possible to determine this case and also provide proper progress indication.
Furthermore we adapt all parameter sources that come with Rally out of the box
and provide guidance for users how to change their own parameter sources.

Relates #763