Store track-related meta-data in results
Rally allows to specify meta-data on track, challenge, task and
operation level. This information is stored for raw metrics records but
not for results. With this commit we ensure that any meta-data are
properly set also on the respective results documents. This allows to
use this information also e.g. in Kibana visualizations.

Relates #771