Store race files always as race.json
With this commit we store the race file always as `race.json`.
Previously Rally had a logic built-in to create a unique file name if
there was a potential for duplicate race files (with the same race
timestamp) but as we're using the race id now for the base directory
this is not possible anymore.

Relates #720
Relates #773