Let the runner determine progress
With this commit we add the ability to have runners determine progress.
Usually the runner is just responsible for executing requests and the
parameter source is determining completion. However, there are cases,
like snapshot/restore or replication where we need to determine progress
based on a request against Elasticsearch.

Relates #789