Change remaining it tests to port 19200 (#825)
The docker tests were using port 9200 because target_hosts was not
specified. Additionally the tests that were using compose were also
leaking a 9200 port on to a default bridge network that other containers
might be using, which causes a collision and failure. This commit ensures
we use a non default network so this collision will not happen.

Closes #824