Manage Elasticsearch nodes with dedicated subcommands
With this commit we introduce three new subcommands to Rally:

* `install`: To install a single Elasticsearch node locally
* `start`: To start an Elasticsearch node that has been previously installed
* `stop`: To stop a running Elasticsearch node

To run a benchmark, users first issue `install`, followed by `start` on all
nodes. Afterwards, the benchmark is run using the `benchmark-only` pipeline.
Finally, the `stop` command is invoked on all nodes to shutdown the cluster.

To ensure that system metrics are stored consistently (i.e. they contain the
same metadata like race id and race timestamp), we expose the race id as a
command line parameter and defer writing any system metrics until the `stop`
command is invoked. We attempt to read race metadata from the Elasticsearch
metrics store for that race id which have been written earlier by the benchmark
and merge the metadata when we write the system metrics.

The current implementation is considered a new experimental addition to the
existing mechanism to manage clusters with the intention to eventually replace
it. The command line interface is specific to Zen discovery and subject to
change as we learn more about its use.

Relates #830
Closes #697