Disable all pylint checks that we want to fix (#857)
This commit will allow us to make progress on the pylint issues we would
like to fix, while allowing users to commit new code that may fail some
of these. What ends up happening is that new code can be merged by other
users, and if precommit is not run in CI, we might accidentally allow a
pylint violation to sneak back in. Removing the pylint check in the
branch/PR that is removing these will allow CI to fail the check so the
user can get feedback if another person has merged code in that violates
the check in the branch/PR.

Relates #838