Log race id on startup (#895)
We have introduced the race id a while ago to uniquely identify a Rally
invocation but we don't write it to logs which makes it harder than
necessary to understand to which race a certain log output belongs to.
With this commit we write the race id Rally's log file upon startup.