Only clear  metrics store on benchmark end (#893)
With this commit we only reset the `metrics_store` field in the driver
instead of completely removing the field from its class. This makes
Rally more resilient in failure scenarios, e.g. when failure happens
upon shutdown of Elasticsearch, which leads to the driver's `#close()`
method being invoked, which in turn attempts to access the already
removed field again.