Add support for a RALLY_HOME env var everywhere (#950)
Previous commits added an override for the new python tests, but this is
not enough as the rally logs and all old tests outputs do not show up in
the proper directory. This commit adds a RALLY_HOME env var that can be
set, which will move everything in rally proper as well as all tests
into that directory, rather than splitting up the data between two
places.