Move CI script to the rally repository (#957)
This script was previously in a repository that did not allow the rally
team to quickly iterate on, and instead relied on a team that is not
responsible for rally related functionality to approve changes
therein. This commit moves the script into the rally repository so that
it can be iterated on by the rally team.