Fix fallback implementation of ML-related runners (#956)
With this commit we fix several issues of the fallback implementation of
ML-related runners. This implementation is only used for older versions
of Elasticsearch which still include the `_xpack` path component and use
the raw transport API of the Elasticsearch client. However, especially
the handling of request parameters did not conform to the API
requirements.

Relates #812