Add tooling to create a track from existing data (#970)
With this commit we introduce a new subcommand `create-track` that allows
users to create a basic Rally track based on data that are already available in
an existing Elasticsearch cluster. Rally will attempt to connect to the target
cluster, extract data and mappings for the specified indices and create a
track that allows to ingest the extract data.

Co-authored-by: Dennis Lawler <dennis.lawler@elastic.co>

Closes #444