Cache source artifacts (#992)
With this commit we add a cache for source artifacts. It can be enabled
by setting `cache = true` in the `source` section of `rally.ini`.
Caching works based on the specified revision. We don't implement an
eviction policy for now but assume that an external process cleans up
cached artifacts.