Improve throughput throttling for bulks (#1100)
With this commit we allow Rally to throttle throughput not only based on
the number of requests per second but also based on any unit that is
supported by the respective parameter source / runner.

We also introduce a feedback loop between the runner and its scheduler,
allowing the scheduler to react to changes in request granularity (e.g.
adapting to changes in bulk size).

Finally, we also introduce an optional feedback loop between the
scheduler and the parameter source but don't use it in Rally's
implementation.

Existing scheduler implementations are deprecated but retrofitted.

Closes #1069