Allow to customize track preparation (#1135)
With this commit we introduce track processors and define two phases in the
track lifecycle where they can hook into:

* `on_after_load_track`: This callback is executed after the track has been
successfully loaded and converted into Rally's domain representation. Rally
defines some out-of-the-box track processors for task filters and test mode
support which will be called in any case.
* `on_prepare_track`: This callback is executed before a load generator will
start executing the benchmark. It can be used to download or generate any
datasets that are needed for the benchmark. Rally defines a default behavior
that downloads and extracts all relevant corpora. That behavior can be
overridden via a custom track processor (i.e. Rally's default behavior is then
*disabled*.).

Like all custom track components, track processors are registered via the
`register` function and we implement a new registration method on the registry
called `register_track_processor` for that, which expects an instance that
implements the two methods above.

Note that track processors cannot share state in between these phases. Some
phases might only be invoked on the coordinating node of a benchmark
(`on_after_load_track`) whereas others will need to be called on each load
generator machine (`on_prepare_track`).

We introduce a new `parameters` dictionary on track as well as on challenge
level, where the challenge properties override the defaults defined track level.
These parameters are only exposed on challenge level and track processors are
expected to operate only on that level. Track processors can determine the
current challenge with the new property `selected_challenge_or_default`.
Mid-term we intend to restrict the current track parameter feature in a way that
all values provided via `--track-params` on the command line must be declared
within the `parameters` block, either on track or on challenge level. That's the
reason why we have called this block already `parameters` although it will only
be used by track processors for now.

We intentionally do not provide user documentation for this feature yet as we
want to gain more experience first before exposing it to a wider user base.

Closes #1066